Simplified the handle_results_click() method down to 4 lines.
By passing in a DBRecord I was able to get rid of the indexing, the
regex and the string slicing. Then I could pass in the DBRecord's fields
to the database search method.

This ended up helping me to fix the problem of most of the lyrics not
appearing in the lyrics box. Now, the lyrics even show up considering
there is still the "missing T" problem.