Removed threading from exact search.
There were too many issues with the threading. It worked in the past,
but I think it was an unsafe operation based on what I've been reading
in the documentation. I went back to a synchronized search to get a MVP
going. I will add a single thread to handle the database queries and
return functionality to the GUI, but multiple threads for queries
doesn't look good. I think I would have to compile my own sqlite for it
to work... I don't want to dive into that right now.